% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{accept_reject_new_seeding_npis}
\alias{accept_reject_new_seeding_npis}
\title{Function to go through to accept or reject proposed parameters for each geoid based
on a geoid specific likelihood.}
\usage{
accept_reject_new_seeding_npis(
  seeding_orig,
  seeding_prop,
  snpi_orig,
  snpi_prop,
  hnpi_orig,
  hnpi_prop,
  hpar_orig,
  hpar_prop,
  orig_lls,
  prop_lls
)
}
\arguments{
\item{seeding_orig}{original seeding data frame (must have column place)}

\item{seeding_prop}{proposal seeding (must have column place)}

\item{snpi_orig}{original npi data frame  (must have column geoid)}

\item{snpi_prop}{proposal npi data frame  (must have column geoid)}

\item{hnpi_orig}{original npi data frame  (must have column geoid)}

\item{hnpi_prop}{proposal npi data frame  (must have column geoid)}

\item{orig_lls}{original ll data frame  (must have column ll and geoid)}

\item{prop_lls}{proposal ll fata frame (must have column ll and geoid)}
}
\value{
a new data frame with the confirmed seedin.
}
\description{
Function to go through to accept or reject proposed parameters for each geoid based
on a geoid specific likelihood.
}
