% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{adaptation_simple}
\alias{adaptation_simple}
\title{Function for adaptive MCMC that generates a new perturbation standard deviation based on the current value, any adaptation parameters, and the past acceptance rate}
\usage{
adaptation_simple(current_sd, adapt, accept_avg, accept_prob)
}
\arguments{
\item{current_sd}{the current perturbation sd}

\item{adapt}{a list of adaptation parameters}

\item{accept_avg}{the average acceptance rate}

\item{accept_prob}{the most recent acceptance probability}
}
\value{
updated perturbation sd
}
\description{
Function for adaptive MCMC that generates a new perturbation standard deviation based on the current value, any adaptation parameters, and the past acceptance rate
}
