% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_perturb_column_hnpi}
\alias{add_perturb_column_hnpi}
\title{Function adds a column to the npi parameter file to record the perturbation standard deviation, initially taken from the config file}
\usage{
add_perturb_column_hnpi(hnpi, intervention_settings)
}
\arguments{
\item{hnpi}{the original npis.}

\item{intervention_settings}{a list of perturbation specifications}
}
\value{
data frame with perturb_sd column added
}
\description{
Function adds a column to the npi parameter file to record the perturbation standard deviation, initially taken from the config file
}
