% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MC_runner_funcs.R
\name{aggregate_and_calc_loc_likelihoods}
\alias{aggregate_and_calc_loc_likelihoods}
\title{Function that performs aggregation and calculates likelihood data across all given locations.}
\usage{
aggregate_and_calc_loc_likelihoods(
  all_locations,
  modeled_outcome,
  obs_nodename,
  config,
  obs,
  ground_truth_data,
  hosp_file,
  hierarchical_stats,
  defined_priors,
  geodata,
  snpi = NULL,
  hnpi = NULL,
  hpar = NULL
)
}
\arguments{
\item{all_locations}{all of the locations to calculate likelihood for}

\item{modeled_outcome}{the hospital data for the simulations}

\item{obs_nodename}{the name of the column containing locations.}

\item{config}{the full configuration setup}

\item{obs}{the full observed data}

\item{ground_truth_data}{the data we are going to compare to aggregated to the right statistic}

\item{hosp_file}{the filename of the hosp file being used (unclear if needed in scope)}

\item{hierarchical_stats}{the hierarchical stats to use}

\item{defined_priors}{information on defined priors.}

\item{geodata}{the geographics data to help with hierarchies}

\item{snpi}{the file with the npi information for seir, only used for heirarchical likelihoods}

\item{hnpi}{the file with the npi information for outcomes, only used for heirarchical likelihoods}

\item{hpar}{data frame of hospitalization parameters, only used for heirarchical likelihoods}
}
\value{
a data frame of likelihood data.
}
\description{
Function that performs aggregation and calculates likelihood data across all given locations.
}
