% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_hierarchical_likadj}
\alias{calc_hierarchical_likadj}
\title{Function to calculate a hierarchical adjustment to the LL
contribution under the assumption that everything comes from
a normal distribution with some variance.}
\usage{
calc_hierarchical_likadj(
  stat,
  infer_frame,
  geodata,
  geo_group_column,
  stat_name_col = "npi_name",
  stat_col = "reduction",
  transform = "none",
  min_sd = 0.1
)
}
\arguments{
\item{stat}{the statistic to calculate the penalty on}

\item{infer_frame}{data frame with the statistics in it}

\item{geodata}{geodata containing geoid from npi fram and the grouping column}

\item{stat_name_col}{column holding stats name...default is npi_name}

\item{stat_col}{column hold the stat}

\item{transform}{how should the data be transformed before calc}

\item{min_sd}{what is the minimum SD to consider. Default is .1}

\item{geo_group_col}{the column to group on}
}
\value{
a data frame with geoids and a per geoid LL adjustment
}
\description{
Function to calculate a hierarchical adjustment to the LL
contribution under the assumption that everything comes from
a normal distribution with some variance.
}
