% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_prior_likadj}
\alias{calc_prior_likadj}
\title{Function to calcualte the likelihood adjustment based on a prior}
\usage{
calc_prior_likadj(params, dist, dist_pars)
}
\arguments{
\item{params}{the parameter values to calculate the likelihood adjust for}

\item{dist}{the distribution to use}

\item{dist_pars}{the parameters of the distribution}
}
\value{
a likelihood sdjustment per param
}
\description{
Function to calcualte the likelihood adjustment based on a prior
}
