% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{coefficientRV}
\alias{coefficientRV}
\title{Reorganized version from
https://rdrr.io/cran/FactoMineR/src/R/coeffRV.R}
\usage{
coefficientRV(X, Y)
}
\description{
Reorganized version from
https://rdrr.io/cran/FactoMineR/src/R/coeffRV.R
}
