% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_cumulative_counts}
\alias{compute_cumulative_counts}
\title{Function to compute cumulative counts across geoids}
\usage{
compute_cumulative_counts(sim_hosp)
}
\arguments{
\item{sim_hosp}{output of ouctomes branching process}
}
\value{
dataframe with the added columns for cumulative counts
}
\description{
Function to compute cumulative counts across geoids
}
