% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_totals}
\alias{compute_totals}
\title{Function to compute cumulative counts across geoids}
\usage{
compute_totals(sim_hosp)
}
\arguments{
\item{sim_hosp}{output of ouctomes branching process}
}
\value{
dataframe with the added rows for all counts
}
\description{
Function to compute cumulative counts across geoids
}
