% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDeduplicationFactors.R
\name{computeDeduplicationFactors}
\alias{computeDeduplicationFactors}
\title{Computes the deduplication factors for each region}
\usage{
computeDeduplicationFactors(
  dupFileName,
  regsFileName,
  postLocPrefix,
  postLocPath
)
}
\arguments{
\item{dupFileName}{The name of the file with the duplicity probabilities. This file is the output of the
\code{deduplication} package.}

\item{regsFileName}{The name of the .csv file defining the regions. It has two columns: \code{ tile, region}. The
first column contains the IDs of each tile in the grid while the second contains the number of a region. This file
is defined by the user and it can be created with any text editor.}

\item{postLocPrefix}{The file name prefix of the files with posterior location probabilities for each device. The
whole file name is composed by a concatenation of \code{prefixName}, \code{_} and \code{deviceID}.The extension of
these files is \code{.dt.csv}}

\item{postLocPath}{The path to the location where the posterior location probabilities are stored.}
}
\value{
A data.table object with the deduplication factors for each region.
}
\description{
Computes the deduplication factors for each region
}
