% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDistrParams.R
\name{computeDistrParams}
\alias{computeDistrParams}
\title{Computes the parameters of the population counts distributions.}
\usage{
computeDistrParams(
  omega,
  popRegFileName,
  pntRateFileName,
  regsFileName = NULL,
  gridFileName = NULL,
  rel_bias = 0,
  cv = 1e-05
)
}
\arguments{
\item{omega}{}

\item{popRegFileName}{The name of the file with the population counts for each region taken from a population
register. It has 2 columns: \code{region, Nreg}.}

\item{pntRateFileName}{The name of the file with the penetration rates for each region. It has 2 columns:
\code{region, pntRate}.}

\item{regsFileName}{The name of the .csv file defining the regions. It has two columns: \code{ tile, region}. The
first column contains the IDs of each tile in the grid while the second contains the number of a region. This file
is defined by the user and it can be created with any text editor. It is required only for the state process
negative binomial distribution.}

\item{gridFileName}{The name of the .csv file with the grid parameters.  It is required only for the state process
negative binomial distribution.}

\item{cv}{The coefficient of variation for the population density of each region. The default value is 0.}

\item{re_bias}{The value of the relative bias for the population density of each region. The default value is 0.}
}
\value{
A data.table object with the following columns  \code{region, omega1, omega2, pnrRate, regionArea_km2, Nreg,
  dedupPntRate, alpha, beta}. If regsFileName and gridFileName are not NULL the result will have 3 more
  columns:\code{region, omega1, omega2, pnrRate, regionArea_km2, Nreg, dedupPntRate, alpha, beta, theta, zeta, Q}. They
  are needed only for the state process negative binomial distribution.
}
\description{
Computes the parameters of the population counts distributions.
}
