% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIntialPopulationStateNB.R
\name{computeInitialPopulationStateNB}
\alias{computeInitialPopulationStateNB}
\title{Computes the distribution of the population count at initial time instants using a state negative binomial
  distribution.}
\usage{
computeInitialPopulationStateNB(nnet, params)
}
\arguments{
\item{nnet}{The random values}

\item{params}{The parameters of the distribution.}
}
\value{
A data.table object
}
\description{
Computes the distribution of the population count at initial time instants using a state negative binomial
  distribution.
}
