% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePopulationT.R
\name{computePopulationT}
\alias{computePopulationT}
\title{Computes population counts at time  instants t >t0.}
\usage{
computePopulationT(
  nt0,
  nnetODFileName,
  zip = TRUE,
  rndVal = FALSE,
  ciprob = NULL,
  method = "ETI"
)
}
\arguments{
\item{nt0}{The population at t0.}

\item{nnetODFileName}{the name of the file where the population moving from one region to another is stored. This is
an output of the \code{aggregation} package.}

\item{zip}{If TRUE the file where where the population moving from one region to another is stored is a zipped csv
file, otherwise it is  simple csv file.}

\item{rndVal}{Controls if the random values generated for each t >t0 are returned or not in the result of this
function. If TRUE, the random values generated according to the corresponding distribution are returned in the
results, if FALSE only the summary statistics for each t>t0 and region are returned.}

\item{ciprob}{Value of probability of the CI (between 0 and 1) to be estimated. If NULL the default value is 0.89.}

\item{method}{The method to compute credible intervals. It could have 2 values, 'ETI' or 'HDI'. The default value is
'ETI.}
}
\value{
A list with one element for each time instant (including t0). Each element of the list is also a list with
  one or two elements, depending on the value of the rndVal parameter. If rndVal is TRUE there are two elements in
  the list corresponding to time instant t. The first one is a data.table object with some descriptive statistics for
  the population count at time t, containing the following columns:\code{region, Mean, Mode, Median, Min, Max, Q1,
  Q3, IQR, SD, CV, CI_LOW, CI_HIGH}. The second one is a data.table object with the random values for population
  count generated for each region, with the following columns: \code{region, iter, NPop}. If rndVal is FALSE the list
  for time instant t contains only the first element previously mentioned. The name of the list element corresponding
  to time instant t is 't' and the name of the two list elements giving the descriptive statistics and random values
  for time t are 'stats' and 'rnd_values'.
}
\description{
Computes the distribution of the population counts for all times instants t > t0.
}
