% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRegionAreas.R
\name{computeRegionAreas}
\alias{computeRegionAreas}
\title{Computes the area of each region}
\usage{
computeRegionAreas(regsFileName, gridFileName)
}
\arguments{
\item{regsFileName}{The name of the .csv file defining the regions. It has two columns: \code{ tile, region}. The
first column contains the IDs of each tile in the grid while the second contains the number of a region. This file
is defined by the user and it can be created with any text editor.}

\item{gridFileName}{The name of the file with the grid parameters. This file could be the one generated by the
simulation software or can be created with any text editor. The grid file generated by the simulation software has
the following columns: \code{Origin X, Origin Y, X Tile Dim, Y Tile Dim, No Tiles X, No Tiles Y}. We are interested
only in the number of rows and columns and the tile size on OX and OY axes. Therefore, the file provided as input to
this function should have at least the following 4 columns: \code{No Tiles X , No Tiles Y, X Tile Dim, Y Tile Dim}.}
}
\value{
A data.table object
}
\description{
Computes the area of each region
}
