% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats.R
\name{computeStats}
\alias{computeStats}
\title{Computes statistics of the population count}
\usage{
computeStats(npop, ciprob, method, by = "region")
}
\arguments{
\item{npop}{A data table with the random values of the population counts generated according to a corresponding
distribution for each region.}

\item{ciprob}{Value of probability of the CI (between 0 and 1) to be estimated. If NULL the default value is 0.89.}

\item{method}{The method to compute credible intervals. It could have 2 values, 'ETI' or 'HDI'. The default value is
'ETI.}
}
\description{
Computes the following values: mean, mode, meadian, min, max, Q1, Q3, IQR , standard deviation,
  coefficient of variation, the value of the credible intreval (LOW, HI).
}
