% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_to_forecast.R
\name{create_cum_death_forecast}
\alias{create_cum_death_forecast}
\title{Creates a merged forecast from a data object and a set of sims.}
\usage{
create_cum_death_forecast(
  sim_data,
  obs_data,
  forecast_date,
  aggregation = "day",
  quants = c(0.01, 0.025, seq(0.05, 0.95, 0.5), 0.975, 0.99),
  weights = NA,
  loc_column = "USPS"
)
}
\arguments{
\item{sim_data}{simulated data.}

\item{obs_data}{the observed data.}

\item{forecast_date}{the date to forecast from (i.e., last date from data)}

\item{aggregation}{the level of aggregation to use.}

\item{quants}{the quantiles to return}

\item{weights}{if not NA, the weights for the mean}

\item{loc_col}{the name of the  location column}
}
\value{
a forecast with columns time (end day), quantile steps_ahead and deaths
}
\description{
Creates a merged forecast from a data object and a set of sims.
}
