% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_to_forecast.R
\name{cum_death_forecast}
\alias{cum_death_forecast}
\title{Functoin that takes the results of an inference run, a date
and a set of cumulative numbers and and results from the simulation.}
\usage{
cum_death_forecast(sim_data, start_date, cum_dat, loc_column)
}
\arguments{
\item{sim_data}{data from the simse to use. already aggregated to correct spatial scale}

\item{start_date}{the date to start from}

\item{cum_dat}{the cumulative data on start date. should have a column to join to sta on and cumDeaths}

\item{loc_column}{whihc column defines location}
}
\value{
a data frame with daily cumulative datra starting from start_date
}
\description{
Functoin that takes the results of an inference run, a date
and a set of cumulative numbers and and results from the simulation.
}
