% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundtruth.R
\name{get_ground_truth}
\alias{get_ground_truth}
\title{Function to load US COVID data from USAfacts}
\usage{
get_ground_truth(
  data_path,
  fips_codes,
  fips_column_name,
  start_date,
  end_date,
  cache = TRUE,
  gt_source = "csse",
  gt_scale = "US county"
)
}
\arguments{
\item{data_path}{Path where to write the data}
}
\description{
Function to load US COVID data from USAfacts
}
