% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundtruth.R
\name{get_ground_truth_file}
\alias{get_ground_truth_file}
\title{Function to load US COVID data from JHUCSSE}
\usage{
get_ground_truth_file(
  data_path,
  cache = TRUE,
  gt_source = "csse",
  gt_scale = "US county"
)
}
\arguments{
\item{data_path}{Path where to write the data}

\item{cache}{logical indicating whether to cache the data (default = TRUE)}

\item{gt_source}{string indicating source of ground truth data. options include "csse" or "usafacts" (default csse)}

\item{gt_scale}{string indicating whether "US county" or "US state"-level data}
}
\description{
Function to load US COVID data from JHUCSSE
}
