% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getStats}
\alias{getStats}
\title{Function for computing statistics over which to compute likelihoods}
\usage{
getStats(df, time_col, var_col, end_date = NULL, stat_list, debug_mode = FALSE)
}
\arguments{
\item{df}{Data frame with data}

\item{time_col}{Name of the column with time}

\item{var_col}{Name of the variable with name of the  column with data to process}

\item{end_date}{Last date to consider}

\item{stat_list}{List with specifications of statistics to compute}
}
\description{
Function for computing statistics over which to compute likelihoods
}
