% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference function for openintro labs}
\usage{
inference(
  y,
  x = NULL,
  est = c("mean", "median", "proportion"),
  success = NULL,
  order = NULL,
  conflevel = 0.95,
  siglevel = 0.05,
  null = NULL,
  alternative = c("less", "greater", "twosided"),
  type = c("ci", "ht"),
  method = c("theoretical", "simulation"),
  simdist = FALSE,
  nsim = 10000,
  seed = NULL,
  sum_stats = TRUE,
  eda_plot = TRUE,
  inf_plot = TRUE,
  inf_lines = TRUE
)
}
\description{
This function is open access, but belongs to the openintro series.
They stopped hosting it within their data for their labs. As such,
I have gone ahead and recreated the variable below. I did not create
this function and do not own the rights to it.
}
