% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MC_runner_funcs.R
\name{initialize_mcmc_first_block}
\alias{initialize_mcmc_first_block}
\title{initialize_mcmc_first_block}
\usage{
initialize_mcmc_first_block(
  run_id,
  block,
  global_prefix,
  chimeric_prefix,
  python_reticulate,
  likelihood_calculation_function,
  is_resume = FALSE
)
}
\arguments{
\item{run_id}{what is the id of this run}

\item{block}{what is the current block}

\item{global_prefix}{the prefix to use for global files}

\item{chimeric_prefix}{the prefix to use for chimeric files}

\item{python_reticulate}{An already initialized copy of python set up to do hospitalization runs}

\item{slot}{what is the current slot number}
}
\description{
initialize_mcmc_first_block
}
