% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iterateAccept}
\alias{iterateAccept}
\title{Function accept proposals}
\usage{
iterateAccept(ll_ref, ll_new)
}
\arguments{
\item{ll_ref}{current accepted likelihood}

\item{ll_new}{likelihood of proposal}
}
\value{
boolean whether to accept the likelihood
}
\description{
Function accept proposals
}
