% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logLikStat}
\alias{logLikStat}
\title{Function for computing statistics over which to compute likelihoods}
\usage{
logLikStat(obs, sim, distr, param, add_one = F)
}
\arguments{
\item{obs}{Vector of observed statistics}

\item{sim}{Vector of simulated statistics}

\item{distr}{Distribution to use for likelihood calculation}

\item{param}{a list opf parameters to the distibution}

\item{add_one}{Whether to add one to simulations to avoid Infs}
}
\description{
Function for computing statistics over which to compute likelihoods
}
