% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{npis_dataframe}
\alias{npis_dataframe}
\title{Create NPIs dataframe for a single location for input to SEIR code}
\usage{
npis_dataframe(
  config,
  random = F,
  geoid = 1,
  offset = 0,
  intervention_multi = 1
)
}
\arguments{
\item{config}{configuration file}

\item{times}{vector of times to use}
}
\value{
a dataframe with npi reduction values by date
}
\description{
Create NPIs dataframe for a single location for input to SEIR code
}
