% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MC_runner_funcs.R
\name{perform_MCMC_step_copies}
\alias{perform_MCMC_step_copies}
\title{Function that performs the necessary file copies the end of an MCMC iteration of
filter_MC.}
\usage{
perform_MCMC_step_copies(
  current_index,
  slot,
  block,
  run_id,
  global_local_prefix,
  gf_prefix,
  global_block_prefix
)
}
\arguments{
\item{current_index}{the current indec in the run}

\item{slot}{what is the current slot numbe}

\item{block}{what is the current block}

\item{run_id}{what is the id of this run}

\item{global_local_prefix}{the prefix to be put on both global and local runs.}

\item{gf_prefix}{the prefix for the directory containing the current globally accepted files.}

\item{global_block_prefix}{prefix that describes this block.}
}
\value{
TRUE if this succeded.
}
\description{
Function that performs the necessary file copies the end of an MCMC iteration of
filter_MC.
}
