% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MC_runner_funcs.R
\name{perform_MCMC_step_copies_chimeric}
\alias{perform_MCMC_step_copies_chimeric}
\title{Function that performs the necessary file copies the end of an MCMC iteration of
filter_MC.}
\usage{
perform_MCMC_step_copies_chimeric(
  current_index,
  slot,
  block,
  run_id,
  chimeric_local_prefix,
  cf_prefix,
  chimeric_block_prefix
)
}
\arguments{
\item{current_index}{the current index in the run}

\item{slot}{what is the current slot number}

\item{block}{what is the current block}

\item{run_id}{what is the id of this run}

\item{chimeric_local_prefix}{the prefix to be put on both chimeric and local runs.}

\item{cf_prefix}{the prefix for the directory containing the current chimericly accepted files.}

\item{chimeric_block_prefix}{prefix that describes this block.}
}
\value{
TRUE if this succeeded.
}
\description{
Function that performs the necessary file copies the end of an MCMC iteration of
filter_MC.
}
