% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{periodAggregate}
\alias{periodAggregate}
\title{Function for applying time aggregation of variables on which to comput likelihoods
Note that bahavior is not consistent when multiples of time nits are passed in.}
\usage{
periodAggregate(
  data,
  dates,
  end_date = NULL,
  period_unit_function,
  period_unit_validator,
  aggregator,
  na.rm = F
)
}
\arguments{
\item{data}{Vector of data to aggregate}

\item{dates}{Vector of dates}

\item{end_date}{Last date to consider}

\item{aggregator}{Function for aggregations}

\item{na.rm}{Remove Nas?}

\item{period_unit}{Unit of period over which to aggregate}

\item{period_k}{Number of time units defining period over which to aggregate}
}
\description{
Function for applying time aggregation of variables on which to comput likelihoods
Note that bahavior is not consistent when multiples of time nits are passed in.
}
