% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{perturb_expand_npis}
\alias{perturb_expand_npis}
\title{Expands the perturb npi into a dataframe with time}
\usage{
perturb_expand_npis(npis, intervention_settings, multi = F)
}
\arguments{
\item{npis}{the npis in "long" version (one value by npi, date and place)}

\item{intervention_settings}{intervention_settings from the config file}
}
\value{
the perturbed npi dataframe
}
\description{
Expands the perturb npi into a dataframe with time
}
