% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perturb_hnpi}
\alias{perturb_hnpi}
\title{Function perturbs an npi parameter file based on
user-specified distributions}
\usage{
perturb_hnpi(hnpi, intervention_settings)
}
\arguments{
\item{hnpi}{the original npis.}

\item{intervention_settings}{a list of perturbation specificationss}
}
\value{
a perturbed data frame
}
\description{
Function perturbs an npi parameter file based on
user-specified distributions
}
