% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perturb_hnpi_from_file}
\alias{perturb_hnpi_from_file}
\title{Function perturbs an npi parameter file based on the current perturbation standard deviation in the file, and also can update the perturbation value if the adaptive mode is turned on}
\usage{
perturb_hnpi_from_file(hnpi, intervention_settings, llik)
}
\arguments{
\item{hnpi}{the original npis.}

\item{intervention_settings}{a list of perturbation specifications}

\item{llik}{log likelihood values}
}
\value{
a perturbed data frame
}
\description{
Function perturbs an npi parameter file based on the current perturbation standard deviation in the file, and also can update the perturbation value if the adaptive mode is turned on
}
