% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perturb_hpar}
\alias{perturb_hpar}
\title{Fucction perturbs an outcomes parameter file based on
user-specified distributions}
\usage{
perturb_hpar(hpar, intervention_settings)
}
\arguments{
\item{hpar}{the original hospitalization (outcomes) parameters.}

\item{intervention_settings}{a list of perturbation specifications}
}
\value{
a perturbed data frame
}
\description{
Fucction perturbs an outcomes parameter file based on
user-specified distributions
}
