% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perturb_snpi}
\alias{perturb_snpi}
\title{Function perturbs an npi parameter file based on
user-specified distributions}
\usage{
perturb_snpi(snpi, intervention_settings)
}
\arguments{
\item{snpi}{the original npis.}

\item{intervention_settings}{a list of perturbation specifications}
}
\value{
a perturbed data frame
}
\description{
Function perturbs an npi parameter file based on
user-specified distributions
}
