% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNnetInitial.R
\name{readNnetInitial}
\alias{readNnetInitial}
\title{Reads the number of individuals detected by network at initial time.}
\usage{
readNnetInitial(nnetFileName)
}
\arguments{
\item{nnetFileName}{The file name of the random values generated by the \code{aggregation} package for the number of
individuals detected by network for each region and each time instant.}
}
\value{
A data.table object with two columns: \code{region, N}
}
\description{
Reads the number of individuals detected by network at initial time instant, as they are generated by
  the \code{aggregation} package.
}
