\name{show,inference-method}
\alias{show,-method}
\alias{show}
\title{Show/print inference object.}
\description{Show/print \code{inference} object.}
\details{\code{show} method for objects made using the \code{infer} function.}
\alias{print.inference}
\value{Nothing.}
\author{Vinh Nguyen}
\arguments{\item{object}{\code{inference} object.}
}

