% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{simulate_multi_epi}
\alias{simulate_multi_epi}
\title{Fill in the epi curve from a baseline state given current parameters}
\usage{
simulate_multi_epi(
  times,
  seedings,
  S0s,
  R0,
  gamma,
  sigma,
  N,
  mob,
  pa = 0.5,
  beta_mults = matrix(rep(1, N * length(times)), ncol = N),
  alpha = 1
)
}
\arguments{
\item{times}{the times to run the epidemic on}

\item{seedings}{the seeding to use}

\item{S0s}{the initial number susceptibles}

\item{gamma}{the time between compartments}

\item{sigma}{the incubation period/latent period}

\item{N}{the number of nodes}

\item{mob}{mobility matrix N x N}

\item{pa}{proportion of time away}

\item{beta_mults}{multipliers to capture the impact of interventions}

\item{alpha}{defaults to 1}

\item{R0s}{the reproductive number}
}
\value{
the epimic states
}
\description{
TODO: Modify to just call python code
}
