% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{simulate_single_epi}
\alias{simulate_single_epi}
\title{Fill in the epi curve from a baseline state given current parameters}
\usage{
simulate_single_epi(
  times,
  seeding,
  S0,
  R0,
  gamma,
  sigma,
  beta_mults = rep(1, length(times)),
  alpha = 1
)
}
\arguments{
\item{times}{the times to run the epidemic on}

\item{seeding}{the seeding to use}

\item{S0}{the initial number susceptible}

\item{R0}{the reproductive number}

\item{gamma}{the time between compartments}

\item{sigma}{the incubation period/latent period}

\item{beta_mults}{multipliers to capture the impact of interventions}

\item{alpha}{defaults to 1}
}
\value{
the epimic states
}
\description{
TODO: Modify to just call python code
}
