% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{single_hosp_run}
\alias{single_hosp_run}
\title{Run hospitalizations and deaths for single epi curve
Functions taken from the hospitalization package}
\usage{
single_hosp_run(epi, config)
}
\arguments{
\item{epi}{SEIR output}

\item{config}{configuration file}
}
\value{
the epimic states
}
\description{
Run hospitalizations and deaths for single epi curve
Functions taken from the hospitalization package
}
