% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{single_loc_inference_test}
\alias{single_loc_inference_test}
\title{Function that does a rapid testing of the inference procedures on a single
time series.}
\usage{
single_loc_inference_test(
  to_fit,
  S0,
  seeding,
  config,
  date_bounds,
  n_slots,
  ncores,
  npi_file,
  seeding_file,
  loglik_file,
  epi_dir
)
}
\arguments{
\item{seeding}{the initial seeding}

\item{config}{the config file with inference info.}

\item{date_bounds}{acceptable bounds for the seeding dates}

\item{n_slots}{the number of slots to run, MCMC iterations per slot are in the config file}

\item{ncores}{the number of cores to run}

\item{npi_file}{file to write NPI samples to}

\item{seeding_file}{file to write seeding samples to}

\item{loglik_file}{file to write neg log-loglik samples to}

\item{epi_dir}{file to write simulated trajectory samples to}
}
\value{
inference run results on this particular area
}
\description{
Function that does a rapid testing of the inference procedures on a single
time series.
}
