% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{synthetic_data}
\alias{synthetic_data}
\title{Create synthetic data to run inference on
Runs an SEIR simulation and applies hospitalization and death on it}
\usage{
synthetic_data(S0, seeding, config)
}
\arguments{
\item{S0}{population size}

\item{config}{configuration file}
}
\value{
the synthetic data
}
\description{
Create synthetic data to run inference on
Runs an SEIR simulation and applies hospitalization and death on it
}
