% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceTest.R
\name{synthetic_data_multi}
\alias{synthetic_data_multi}
\title{Create synthetic data to run inference on
Runs an SEIR simulation and applies hospitalization and death on it}
\usage{
synthetic_data_multi(S0s, seedings, mob, config, offsets, interventions_multi)
}
\arguments{
\item{S0s}{population sizes}

\item{seedings}{seeding dataframe}

\item{config}{configuration file}
}
\value{
the synthetic data
}
\description{
Create synthetic data to run inference on
Runs an SEIR simulation and applies hospitalization and death on it
}
