% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDistrParams.R
\name{computeDistrParams}
\alias{computeDistrParams}
\title{Computes the parameters of the population counts distributions.}
\usage{
computeDistrParams(omega, popRegFileName, pntRateFileName)
}
\arguments{
\item{omega}{}

\item{popRegFileName}{}

\item{pntRateFileName}{}
}
\value{
A data.table object
}
\description{
Computes the parameters of the population counts distributions.
}
