% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\docType{package}
\name{inference}
\alias{inference}
\title{inference: A package for computing the distribution of the number of individuals in the target population conditioned on the number of
individuals detected by MNO.}
\description{
This package contains functions to compute the distribution of the number of individuals in the target population
conditioned on the number of individuals detected by MNO. For an example on how to use this package please read
\link{example} or \link{exampleAPI}.
}
