% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perturb_seeding}
\alias{perturb_seeding}
\title{Fuction perturbs a seeding file based on a normal
proposal on the start date and
a poisson on the number of cases.}
\usage{
perturb_seeding(seeding, sd, date_bounds)
}
\arguments{
\item{seeding}{the original seeding}

\item{sd}{the standard deviation of the posson}
}
\value{
a pertubed data frame
}
\description{
Fuction perturbs a seeding file based on a normal
proposal on the start date and
a poisson on the number of cases.
}
