% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{perturb_seeding}
\alias{perturb_seeding}
\title{Function perturbs a seeding file based on a normal
proposal on the start date and
a poisson on the number of cases.}
\usage{
perturb_seeding(
  seeding,
  date_sd,
  date_bounds,
  amount_sd = 1,
  continuous = FALSE
)
}
\arguments{
\item{seeding}{the original seeding}

\item{date_sd}{the standard deviation parameter of the normal distribution used to perturb date}

\item{amount_sd}{the standard deviation parameter of the normal distribution used to perturb amount}

\item{continuous}{Whether the seeding is passed to a continuous model or not}
}
\value{
a perturbed data frame
}
\description{
Function perturbs a seeding file based on a normal
proposal on the start date and
a poisson on the number of cases.
}
