\name{show,inference-method}
\alias{show,-method}
\alias{show}
\title{Show/print \linkS4class{inference} object.}
\description{Show/print \code{\linkS4class{inference}} object.}
\details{\code{\link[methods]{show}} method for objects made using the \code{\link{infer}} function.}
\alias{print.inference}
\value{Nothing.}
\author{Vinh Nguyen}
\arguments{\item{object}{\code{\linkS4class{inference}} object.}
}

